feat: add test utilities and the infrastructure system mapper

+++ momentum-acquisition/momentum-acquisition-domain/src/main/java/dev/zalaya/momentum/acquisition/domain/model/Process.java
@@ -9,7 +9,7 @@ public class Process {
     String name;
     String path;
     Double cpu;
-    Double memory;
+    Long memory;
     Integer threads;
 
 }
